Fixed Bug related to nagios plugin output.
When using relative values and --unit with check_jmx4perl, then the unit
for the value is used for the base value, regardless of the normalization
is used for the base value which can differ from the conversion of the
value itself. You can use %w in the placedholder for specifying the base
value's unit.